harden SHM ring, monitor reader, VMM teardown; scrub internal refs
shm_ring: eliminate &mut [u8] aliasing UB via raw pointer volatile
ops. cap guest-controlled msg.length before allocation. use
wrapping_sub modular distance with used>capacity invariant guard on
drain paths. guard shm_drain against undersized buffers.

reader: checked_add for pa+offset arithmetic. bound read_bytes and
scalar ops by resolved region remainder. aarch64 checked_sub for
descriptors below DRAM_START.

vmm/mod: .unwrap() → .context()? on get_host_address. defensive
Ctrl-A byte reorder. shm_size >= HEADER_SIZE guard.

flock: u16 cast in octal decoder prevents overflow on \400+. delete
unused needle_from_fd. enumerate 4 flock surfaces in module doc.

cache: reject pid <= 0 in orphan sweep. migrate atomic_swap_dirs to
rustix::fs::renameat_with.

btf_offsets: checked_add for ELF section header bounds.

fetch: connect_timeout(10s) on shared client. warn on dropped
releases.json rows.

stats: delete compare_rows/group_and_average wrappers. wire
wake_latency_tail_ratio and iterations_per_worker into DataFrame.
rename Dimension::Commit→ProjectCommit, Source→RunSource.

cli: gate source_vmlinux_stat on is_local_source. warn on
KTSTR_KERNEL_PARALLELISM parse failure. accept uppercase hex in
looks_like_literal_hash. drop group_and_average re-export.

macros: remove sockets/min_sockets aliases. add helper unit tests.

docs: lead with cargo ktstr test. align error strings to source.
update model size to 2.55 GiB. remove backward-compat notes.

tests: deduplicate scheduler consts. extract TOPO_1L_4C_1T. add
expect_err comments. remove BPF line-number references.

scrub internal task numbers and test-plan codes from comments.

toolchain: MSRV 1.94.1, minimal profile.

+++ src/host_state.rs
@@ -1520,7 +1520,7 @@ mod tests {
         let snap = capture();
         assert!(!snap.threads.is_empty());
         // Every captured thread carries non-ephemeral identity.
-        let pid = std::process::id() as u32;
+        let pid = std::process::id();
         let self_threads: Vec<_> = snap.threads.iter().filter(|t| t.tgid == pid).collect();
         assert!(!self_threads.is_empty(), "own tgid missing from capture");
     }
